hyphen_alchemy_4/app/requirements.txt
bcrypt
openpyxl
pillow
pycparser
pydantic
pydantic_core
PyMySQL
mysqlclient
python-dateutil
pytz
reportlab
XlsxWriter
fastapi
mysql-connector-python
passlib
psycopg2
uvicorn
Werkzeug
openai
langchain_openai
bs4
psutil
Pyjwt
vertica_python
python-dotenv
cryptography
pydantic[email]
cx_oracle
sqlalchemy
psycopg2-binary
vertica-python
pandas
pydantic-settings
python-multipart
redis
aiomysql
asyncpg
tenacity
aioredis
gunicorn
boto3
